refactor(UsersRoutes): move component to its own directory
